requirements.txt
Flask
Flask-Mail
python-dotenv
sqlite3
logging
